set common base class with __repr__ and __str__
