+++ visonic/devices.py
@@ -1,38 +1,13 @@
-from dataclasses import dataclass, field
-import inspect
-import json
+from dataclasses import dataclass
+
+from .classes import BaseClass
 
 
 @dataclass
-class BaseDevice:
+class BaseDevice(BaseClass):
     """Base class definition of a device in the alarm system."""
 
     _device: dict
-
-    def __repr__(self):
-        r = ""
-        attrs = inspect.getmembers(self, lambda a: not (inspect.isroutine(a)))
-        for i, a in enumerate([attr for attr in attrs if self._is_property(attr)]):
-            if i:
-                r = r + ", "
-            r = r + f"{a[0]} = {getattr(self, a[0])}"
-        return f"{type(self).__name__}({r})"
-
-    def __str__(self):
-        r = {}
-        attrs = inspect.getmembers(self, lambda a: not (inspect.isroutine(a)))
-        for a in [attr for attr in attrs if self._is_property(attr)]:
-            r[a[0]] = getattr(self, a[0])
-        return f"{str(type(self))}: {r}"
-
-    def _is_property(self, attr) -> bool:
-        if not (
-            attr[0].startswith("__")
-            and attr[0].endswith("__")
-            or attr[0].startswith("_")
-        ):
-            return True
-        return False
 
     # Device properties
     @property
